fix: HASSU-870 ei vaadita kiinteistönomistajiem hakua aloituskuulutuksen uudelleenkuulutuksessa, jos tiedotaKiinteistonomistajia on false

+++ backend/src/handler/tila/aloitusKuulutusTilaManager.ts
@@ -146,7 +146,9 @@ class AloitusKuulutusTilaManager extends KuulutusTilaManager<AloitusKuulutus, Al
       throw new IllegalAineistoStateError();
     }
     const suomifiViestitEnabled = await parameters.isSuomiFiViestitIntegrationEnabled();
-    if (suomifiViestitEnabled && !projekti.omistajahaku?.status) {
+    // Aloituskuulutuksen uudelleenkuulutustilanteessa ei vaadita kiinteistönomistajia jos ei haluta tiedottaa ennen ko. toiminnallisuutta aloituskuulutettuja
+    const skipKiinteistoRequirement = vaihe.uudelleenKuulutus?.tiedotaKiinteistonomistajia === false;
+    if (suomifiViestitEnabled && !projekti.omistajahaku?.status && !skipKiinteistoRequirement) {
       const msg = "Kiinteistönomistajia ei ole haettu ennen aloituskuulutuksen hyväksyntää";
       log.error(msg);
       throw new IllegalArgumentError(msg);
@@ -321,11 +323,13 @@ class AloitusKuulutusTilaManager extends KuulutusTilaManager<AloitusKuulutus, Al
           filePathInProjekti: pdfs.aloituskuulutusIlmoitusPDFPath,
           reason: "Aloituskuulutus rejected",
         });
-        await fileService.deleteYllapitoFileFromProjekti({
-          oid,
-          filePathInProjekti: pdfs.aloituskuulutusIlmoitusKiinteistonOmistajallePDFPath,
-          reason: "Aloituskuulutus rejected",
-        });
+        if (pdfs.aloituskuulutusIlmoitusKiinteistonOmistajallePDFPath) {
+          await fileService.deleteYllapitoFileFromProjekti({
+            oid,
+            filePathInProjekti: pdfs.aloituskuulutusIlmoitusKiinteistonOmistajallePDFPath,
+            reason: "Aloituskuulutus rejected",
+          });
+        }
       }
     }
   }
